+++ src/entities/concert/ui/ConcertInfoTab.tsx
@@ -1,6 +1,12 @@
 import { useEffect, useState } from "react";
 import { getArtistInfo, Artist } from "../api/getArtistInfo";
 import { getConcertCulture, ConcertCulture } from "../api/getConcertCulture";
+import { getSchedule, Schedule } from "../api/getSchedule";
+import {
+  getConcertRequiredInfo,
+  ConcertRequired,
+} from "../api/getConcertRequiredInfo";
+import { getMd, Md } from "../api/getMd";
 import ArtistTabPanel from "./ArtistTabPanel";
 import ConcertTabPanel from "./ConcertTabPanel";
 import EmptyConcertInfoTabPanel from "./EmptyConcertInfoTabPanel";
@@ -19,8 +25,14 @@ interface ConcertInfoTabProps {
 
 function ConcertInfoTab({ concertId, ticketUrl }: ConcertInfoTabProps) {
   const [selectedTab, setSelectedTab] = useState("artist");
+
   const [artist, setArtist] = useState<Artist | null>(null);
   const [ConcertCulture, setConcertCulture] = useState<ConcertCulture[]>([]);
+  const [schedules, setSchedule] = useState<Schedule[] | null>(null);
+  const [concertRequiredInfo, setConcertRequiredInfo] = useState<
+    ConcertRequired[] | null
+  >(null);
+  const [mds, setMd] = useState<Md[] | null>(null);
 
   useEffect(() => {
     if (!concertId || isNaN(concertId)) {
@@ -49,6 +61,48 @@ function ConcertInfoTab({ concertId, ticketUrl }: ConcertInfoTabProps) {
     };
 
     fetchConcertCulture();
+  }, [concertId]);
+
+  useEffect(() => {
+    const fetchSchedule = async () => {
+      try {
+        const data = await getSchedule(concertId);
+        setSchedule(data);
+      } catch (error) {
+        console.error("특정 콘서트 일정 목록 조회 API 호출 실패:", error);
+        setSchedule([]);
+      }
+    };
+
+    fetchSchedule();
+  }, [concertId]);
+
+  useEffect(() => {
+    const fetchConcertRequiredInfo = async () => {
+      try {
+        const data = await getConcertRequiredInfo(concertId);
+        setConcertRequiredInfo(data);
+      } catch (error) {
+        console.error("특정 콘서트 필수 정보 목록 조회 API 호출 실패:", error);
+        setConcertRequiredInfo([]);
+      }
+    };
+
+    fetchConcertRequiredInfo();
+  }, [concertId]);
+
+  useEffect(() => {
+    const fetchMds = async () => {
+      try {
+        const data = await getMd(concertId);
+        setMd(data);
+      } catch (error) {
+        console.error("특정 콘서트의 MD 목록 조회 API 호출 실패:", error);
+        setMd([]);
+      }
+    };
+
+    fetchMds();
   }, [concertId]);
 
   return (
@@ -132,8 +186,19 @@ function ConcertInfoTab({ concertId, ticketUrl }: ConcertInfoTabProps) {
           )}
         </TabPanel>
         <TabPanel value="concert" className="p-0">
-          <ConcertTabPanel concertId={concertId} ticketUrl={ticketUrl} />
-          {/* <EmptyConcertInfoTabPanel text={"콘서트 정보"} /> */}
+          {(!schedules || schedules.length === 0) &&
+          (!concertRequiredInfo || concertRequiredInfo.length === 0) &&
+          (!mds || mds.length === 0) ? (
+            <EmptyConcertInfoTabPanel text={"콘서트 정보"} />
+          ) : (
+            <ConcertTabPanel
+              concertId={concertId}
+              ticketUrl={ticketUrl}
+              schedules={schedules}
+              concertRequiredInfo={concertRequiredInfo}
+              mds={mds}
+            />
+          )}
         </TabPanel>
         <TabPanel value="setlist">셋리스트 내용</TabPanel>
       </TabsBody>
