+++ src/features/concert/ui/ConcertSetting.tsx
@@ -7,8 +7,8 @@ import Tab from "@mui/material/Tab";
 import TabContext from "@mui/lab/TabContext";
 import TabList from "@mui/lab/TabList";
 import TabPanel from "@mui/lab/TabPanel";
-import ConcertSchedulePanel from "./ConcertSchedulePanel";
 import EmptyConcertSchedulePanel from "./EmptyConcertSchedulePanel";
+import ScheduleInfo from "../../../entities/concert/ui/ScheduleInfo";
 
 function ConcertSetting() {
   const [tabValue, setTabValue] = useState("1");
@@ -145,7 +145,7 @@ function ConcertSetting() {
               <br />
               확인해 보세요!
             </div>
-            <ConcertSchedulePanel />
+            <ScheduleInfo />
             {/* <EmptyConcertSchedulePanel /> */}
           </TabPanel>
           <TabPanel
